[level 2] Title: k진수에서 소수 개수 구하기, Time: 0.27 ms, Memory: 81.2 MB -BaekjoonHub

+++ 프로그래머스/2/92335. k진수에서 소수 개수 구하기/k진수에서 소수 개수 구하기.java
@@ -2,7 +2,7 @@ import java.util.*;
 
 class Solution {
     public int solution(int n, int k) {
-        String trans = Long.toString(n, k);
+        String trans = Integer.toString(n, k);
         String[] str = trans.split("[0]+");
         
         int cnt = 0;
